% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clin_info.R
\name{add_interpretation_columns}
\alias{add_interpretation_columns}
\title{Add Clinical interpretation columns}
\usage{
add_interpretation_columns(snv)
}
\arguments{
\item{snv}{}
}
\description{
Add Clinical interpretation columns
}
