% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_QC.R
\name{add_strand_ratios}
\alias{add_strand_ratios}
\title{Add columns containing forward reverse strand coverage info}
\usage{
add_strand_ratios(snv)
}
\arguments{
\item{snv}{}
}
\description{
Add columns containing forward reverse strand coverage info
}
