% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathoWin_Annotation.R
\name{diagnose_D_F_columns_snv}
\alias{diagnose_D_F_columns_snv}
\title{\subsection{Prepare Diagnosis column in German and French to be copy/pasted to Pathowin}{

SNV annotation
}}
\usage{
diagnose_D_F_columns_snv(variant_tbl)
}
\arguments{
\item{variant_tbl}{}
}
\description{
\subsection{Prepare Diagnosis column in German and French to be copy/pasted to Pathowin}{

SNV annotation
}
}
