% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_QC.R
\name{display_std_metrics}
\alias{display_std_metrics}
\title{Select few columns for metrics inspection}
\usage{
display_std_metrics(snv)
}
\arguments{
\item{snv}{}
}
\description{
Select few columns for metrics inspection
}
