% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_QC.R
\name{display_variants}
\alias{display_variants}
\title{Select few columns for variant inspection}
\usage{
display_variants(snv)
}
\arguments{
\item{snv}{}
}
\description{
Select few columns for variant inspection
}
