% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clin_info.R
\name{read_cancerHotspot}
\alias{read_cancerHotspot}
\title{read in annotation FP cancerHotspot}
\usage{
read_cancerHotspot(cancerHotspot_path)
}
\arguments{
\item{cancerHotspot_path}{}
}
\description{
read in annotation FP cancerHotspot
}
