% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_filepaths.R
\name{reporting_filepaths}
\alias{reporting_filepaths}
\title{Generate list of filepaths for reporting folder}
\usage{
reporting_filepaths(output_path)
}
\arguments{
\item{output_path}{}
}
\description{
Generate list of filepaths for reporting folder
}
